requirements.txt
firebase-admin>=5.0.0
mysql-connector-python>=8.0.0
python-dotenv>=0.19.0
